requirements.txt
#discord.py-self
git+https://github.com/dolfies/discord.py-self@renamed
python-dotenv
google-generativeai
